extracts buyer select
